# Reduce memory allocations in JavaScriptParser