With this PR we are pretty much getting all the allocations from AST + strings. I do have a StringSegment/TextSpan branch somewhere though...

* allow clearing of and supplying initial data directly to ArrayList
* make ParsedParameters a struct
* pool data structures used by ParseBinaryExpression, ParseLexicalBinding and ParsePattern

Also Jint tests pass.