@lahma 2022-08-15
## Esprima.Benchmark.FileParsingBenchmark

| **Diff**|Method|FileName|Mean|Error|Gen 0|Gen 1|Allocated|
|------- |-------|-------|-------:|-------|-------:|-------:|-------:|
| Old |ParseProgram|angular-1.2.5|10.554 ms|0.0166 ms|421.8750|156.2500|6,892 KB|
| **New** |	|	| **10.617 ms (+1%)** | **0.0244 ms** | **359.3750 (-15%)** | **156.2500 (0%)** | **6,010 KB (-13%)** |
| Old |ParseProgram|backbone-1.1.0|1.388 ms|0.0021 ms|62.5000|25.3906|1,036 KB|
| **New** |	|	| **1.388 ms (0%)** | **0.0101 ms** | **54.6875 (-13%)** | **27.3438 (+8%)** | **913 KB (-12%)** |
| Old |ParseProgram|jquery-1.9.1|8.611 ms|0.0190 ms|359.3750|171.8750|5,976 KB|
| **New** |	|	| **8.396 ms (-2%)** | **0.0208 ms** | **312.5000 (-13%)** | **156.2500 (-9%)** | **5,144 KB (-14%)** |
| Old |ParseProgram|jquery.mobile-1.4.2|13.704 ms|0.0360 ms|531.2500|265.6250|8,901 KB|
| **New** |	|	| **13.033 ms (-5%)** | **0.0201 ms** | **468.7500 (-12%)** | **218.7500 (-18%)** | **7,824 KB (-12%)** |
| Old |ParseProgram|mootools-1.4.5|6.542 ms|0.0114 ms|289.0625|109.3750|4,812 KB|
| **New** |	|	| **6.629 ms (+1%)** | **0.0131 ms** | **250.0000 (-14%)** | **117.1875 (+7%)** | **4,182 KB (-13%)** |
| Old |ParseProgram|underscore-1.5.2|1.174 ms|0.0018 ms|56.6406|19.5313|943 KB|
| **New** |	|	| **1.176 ms (0%)** | **0.0048 ms** | **46.8750 (-17%)** | **19.5313 (0%)** | **792 KB (-16%)** |
| Old |ParseProgram|yui-3.12.0|6.177 ms|0.0132 ms|257.8125|125.0000|4,301 KB|
| **New** |	|	| **6.062 ms (-2%)** | **0.0100 ms** | **234.3750 (-9%)** | **109.3750 (-13%)** | **3,850 KB (-10%)** |