@sebastienros 2022-08-15
Was wondering why you didn't share tables anymore ;)