@lahma 2022-08-15
This is the current top allocation percentages when parsing `jquery.mobile-1.4.2.js` 100 times with this PR code. 

![image](https://user-images.githubusercontent.com/171892/184680701-f8e037d4-087d-4a30-a799-b80aef09154f.png)