@adams85 2022-08-16
Nice improvements again, you're squeezing those allocations like hell! 😄 

BTW, [this benchmark](https://github.com/sebastienros/esprima-dotnet/pull/313#issuecomment-1215379293) of yours showing that `string`s and `Identifier`s are on the top of the list gave me an idea how we might push this even further:

Let's suppose we have the following input:
```js
let abc;
abc = 0;
```

We declare something, then later we reference that using its name. Pretty common stuff. When we parse this, we end up with a bunch of identical `Identifier`s (well, not completely identical ones because of the metadata like location, range, etc.) However, the `Name` property of these identifiers will be exactly the same strings. I mean they'll contain the same characters. They won't be the same *reference* (apart from some special cases covered by `ParserExtensions.Slice`) currently though. But they could be and this is where I see a room for improvement.

It looks possible to introduce some kind of a string pooling by means of which we could make sure that `Scanner` won't allocate the same string multiple times. We may borrow some ideas from [here](https://docs.microsoft.com/en-us/windows/communitytoolkit/high-performance/stringpool) as it looks to do exactly what I'm talking about.