@lahma 2022-08-16
Thanks for the review, I'll try to address these things as time permits. Workdays with family can be a bit... crowded...

> BTW, [this benchmark](https://github.com/sebastienros/esprima-dotnet/pull/313#issuecomment-1215379293) of yours showing that `string` and `Identifier` are sitting on the top of the list gave me an idea how we might push this even further:
> 
> Let's suppose we have the following input:
> 
> ```js
> let abc;
> abc = 0;
> ```
> 
> We declare something, then later we reference that using its name. Pretty common stuff. When we parse this, we end up with an AST containing two identical `Identifier`s (well, not completely identical because of metadata like location, range, etc.) However, the `Name` property of these identifiers will be exactly the same strings. I mean they'll contain the same characters. They won't currently be the same _instance_ (apart from some special cases covered by `ParserExtensions.Slice`) though. But they could be and this is where I see a room for improvement.
> 
> It looks possible to introduce some kind of a string pooling by means of which we could make sure that `Scanner` won't allocate the same string multiple times. We may borrow some ideas from [here](https://docs.microsoft.com/en-us/windows/communitytoolkit/high-performance/stringpool) as it looks to do exactly what I'm talking about.

Yes some time ago I was wondering about `Identifier` pooling, which basically dies down when location and range are introduced, like you've said, Jint needs them anyways always for proper problem reporting. I think pooling of string identifiers is a competing solution to `StringSegment/TextSpan` - this requires benchmarking and pros and cons.

I discussed this with Sebastien and ideally we wouldn't introduce a new library for this. Copying existing (MIT or similar) code is OK with attribution as probably headers on top of file.