@lahma 2022-08-16
Did a review round fixes, current stats:

## Esprima.Benchmark.FileParsingBenchmark

| **Diff**|Method|FileName|Mean|Error|Gen 0|Gen 1|Allocated|
|------- |-------|-------|-------:|-------|-------:|-------:|-------:|
| Old |ParseProgram|angular-1.2.5|10.635 ms|0.0835 ms|421.8750|156.2500|6,892 KB|
| **New** |	|	| **10.467 ms (-2%)** | **0.0665 ms** | **343.7500 (-19%)** | **156.2500 (0%)** | **5,751 KB (-17%)** |
| Old |ParseProgram|backbone-1.1.0|1.416 ms|0.0042 ms|62.5000|25.3906|1,036 KB|
| **New** |	|	| **1.381 ms (-2%)** | **0.0113 ms** | **52.7344 (-16%)** | **21.4844 (-15%)** | **888 KB (-14%)** |
| Old |ParseProgram|jquery-1.9.1|8.708 ms|0.0183 ms|359.3750|171.8750|5,976 KB|
| **New** |	|	| **8.263 ms (-5%)** | **0.0391 ms** | **296.8750 (-17%)** | **140.6250 (-18%)** | **4,995 KB (-16%)** |
| Old |ParseProgram|jquery.mobile-1.4.2|13.556 ms|0.0296 ms|531.2500|265.6250|8,901 KB|
| **New** |	|	| **13.379 ms (-1%)** | **0.0428 ms** | **468.7500 (-12%)** | **234.3750 (-12%)** | **7,678 KB (-14%)** |
| Old |ParseProgram|mootools-1.4.5|6.716 ms|0.0149 ms|289.0625|109.3750|4,812 KB|
| **New** |	|	| **6.479 ms (-4%)** | **0.0172 ms** | **242.1875 (-16%)** | **85.9375 (-21%)** | **4,033 KB (-16%)** |
| Old |ParseProgram|underscore-1.5.2|1.207 ms|0.0038 ms|56.6406|19.5313|943 KB|
| **New** |	|	| **1.150 ms (-5%)** | **0.0018 ms** | **44.9219 (-21%)** | **17.5781 (-10%)** | **748 KB (-21%)** |
| Old |ParseProgram|yui-3.12.0|6.143 ms|0.0115 ms|257.8125|125.0000|4,301 KB|
| **New** |	|	| **5.986 ms (-3%)** | **0.0313 ms** | **226.5625 (-12%)** | **85.9375 (-31%)** | **3,792 KB (-12%)** |