@adams85 2022-08-16
> Thanks for the review, I'll try to address these things as time permits. Workdays with family can be a bit... crowded...

No problem. Hope you don't mind, I did some further corrections (see https://github.com/sebastienros/esprima-dotnet/pull/313/commits/309a481890c3fb6fff4782ffcc03456477c70246). It was faster for me this way than doing another review iteration. Check out these changes and if you agree, we can go ahead with merging.

> Yes some time ago I was wondering about Identifier pooling, which basically dies down when location and range are introduced, like you've said, Jint needs them anyways always for proper problem reporting. I think pooling of string identifiers is a competing solution to StringSegment/TextSpan - this requires benchmarking and pros and cons.

I discussed this with Sebastien and ideally we wouldn't introduce a new library for this. Copying existing (MIT or similar) code is OK with attribution as probably headers on top of file.

Fully agreed. I'll explore the string pooling approach and will get back to you in case the results look promising.