@lahma 2022-08-17
Thanks for the help! A couple kilobyte increase in some benchmarks after latest commit but really insignificant.